# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DrugRequestStatus =
  (typeof DrugRequestStatus)[keyof typeof DrugRequestStatus];

export const DrugRequestStatus = {
  pending: "pending",
  responded: "responded",
} as const;

export interface DrugRequest {
  id: string;
  userId: string;
  drugName: string;
  status: DrugRequestStatus;
  createdAt: string;
  respondedAt?: string | null;
  pharmacyName?: string | null;
  pharmacyAddress?: string | null;
  pharmacyPhone?: string | null;
}

export interface CreateDrugRequest {
  userId: string;
  drugName: string;
}

export interface RespondToRequest {
  pharmacyName: string;
  pharmacyAddress: string;
  pharmacyPhone: string;
}

export interface Notification {
  id: string;
  userId: string;
  requestId: string;
  pharmacyName: string;
  pharmacyAddress: string;
  pharmacyPhone: string;
  isLocked: boolean;
  isRead: boolean;
  createdAt: string;
}