# lib/api-zod/src/generated/types/createDrugRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDrugRequest {
  userId: string;
  drugName: string;
}