# lib/api-zod/src/generated/types/drugRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { DrugRequestStatus } from "./drugRequestStatus";

export interface DrugRequest {
  id: string;
  userId: string;
  drugName: string;
  status: DrugRequestStatus;
  createdAt: string;
  respondedAt?: string | null;
  pharmacyName?: string | null;
  pharmacyAddress?: string | null;
  pharmacyPhone?: string | null;
}