# lib/api-zod/src/generated/types/notification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Notification {
  id: string;
  userId: string;
  requestId: string;
  pharmacyName: string;
  pharmacyAddress: string;
  pharmacyPhone: string;
  isLocked: boolean;
  isRead: boolean;
  createdAt: string;
}