# lib/api-zod/src/generated/types/respondToRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface RespondToRequest {
  pharmacyName: string;
  pharmacyAddress: string;
  pharmacyPhone: string;
}